+++ src/main/kotlin/com/github/noahshen/panther/core/BlockHeader.kt
@@ -9,6 +9,6 @@ data class BlockHeader(
         val coinBase: ByteArray,
         val time: DateTime,
         val difficulty: Long,
-        val nonce: Int,
+        val nonce: Long,
         val trxTrieRoot: ByteArray
 )
